pyproject.toml
[project]
name = "sigos-ufcat"
version = "0.1.0"
description = "Sistema de gestão de ordens de serviço"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.2",
    "django-browser-reload>=1.17.0",
    "django-stubs>=5.1.1",
    "django-widget-tweaks>=1.5.0",
    "fontawesomefree>=6.6.0",
    "requests>=2.32.3",
    "faker>=33.1.0",
    "django-htmx>=1.21.0",
    "django-cors-headers>=4.6.0",
    "pillow>=11.1.0",
    "django-simple-history>=3.8.0",
    "django-allauth>=65.3.1",
    "python-dotenv>=1.0.1",
    "psycopg2-binary>=2.9.10",
    "psycopg2>=2.9.10",
    "colorama>=0.4.6",
    "djangorestframework>=3.15.2",
    "dj-database-url>=2.3.0",
    "whitenoise>=6.9.0",
    "gunicorn>=23.0.0",
    "django-heroku>=0.3.1",
]
